Refactor database (bux fix second way)

+++ JavaService/scrapper/src/main/resources/db/migrations/2023-04-03-BT-1_create_client_table.sql
@@ -4,6 +4,6 @@
 
 create table if not exists client
 (
-    chat_id   bigint primary key not null,
+    chat_id   serial primary key not null,
     user_name text               not null
 );
